+++ substreams-macro/src/handler.rs
@@ -23,8 +23,6 @@ pub fn main(_args: TokenStream, item: TokenStream, module_type: ModuleType) -> T
     let mut read_only_stores: Vec<proc_macro2::TokenStream> =
         Vec::with_capacity(input.sig.inputs.len());
     let mut writable_store: proc_macro2::TokenStream = quote! {};
-    let mut forgets: Vec<proc_macro2::TokenStream> =
-        Vec::with_capacity(input.sig.inputs.len());
 
     for i in (&input.sig.inputs).into_iter() {
         match i {
@@ -95,8 +93,7 @@ pub fn main(_args: TokenStream, item: TokenStream, module_type: ModuleType) -> T
                                 let #var_name: #argument_type = substreams::store::Deltas::new(#raw);
                             })
                         } else if input_obj.is_string {
-                            proto_decodings.push(quote! { let #var_name: String = unsafe {String::from_raw_parts(#var_ptr, #var_len, #var_len) }; });
-                            forgets.push(quote!{ unsafe {std::mem::forget(#var_name)}; });
+                            proto_decodings.push(quote! { let #var_name: String = std::mem::ManuallyDrop::new(unsafe {String::from_raw_parts(#var_ptr, #var_len, #var_len)}).to_string(); });
                         } else {
                             proto_decodings.push(quote! { let #var_name: #argument_type = substreams::proto::decode_ptr(#var_ptr, #var_len).unwrap(); })
                         }
@@ -117,7 +114,6 @@ pub fn main(_args: TokenStream, item: TokenStream, module_type: ModuleType) -> T
             input,
             args,
             proto_decodings,
-            forgets,
             read_only_stores,
             writable_store,
         ),
@@ -125,7 +121,6 @@ pub fn main(_args: TokenStream, item: TokenStream, module_type: ModuleType) -> T
             input,
             args,
             proto_decodings,
-            forgets,
             read_only_stores,
             writable_store,
         ),
@@ -268,7 +263,6 @@ fn build_map_handler(
     input: syn::ItemFn,
     collected_args: Vec<proc_macro2::TokenStream>,
     decodings: Vec<proc_macro2::TokenStream>,
-    forgets: Vec<proc_macro2::TokenStream>,
     read_only_stores: Vec<proc_macro2::TokenStream>,
     writable_store: proc_macro2::TokenStream,
 ) -> TokenStream {
@@ -284,7 +278,6 @@ fn build_map_handler(
             #(#read_only_stores)*
             #writable_store
             let result = #body;
-            #(#forgets)*
             result
         };
     };
@@ -307,7 +300,6 @@ fn build_store_handler(
     input: syn::ItemFn,
     collected_args: Vec<proc_macro2::TokenStream>,
     decodings: Vec<proc_macro2::TokenStream>,
-    forgets: Vec<proc_macro2::TokenStream>,
     read_only_stores: Vec<proc_macro2::TokenStream>,
     writable_store: proc_macro2::TokenStream,
 ) -> TokenStream {
@@ -324,7 +316,6 @@ fn build_store_handler(
             #(#read_only_stores)*
             #writable_store
             let result = #body;
-            #(#forgets)*
             result
         }
     };